correciones de usuario

+++ src/main/java/pe/edu/upc/aaw/demoproyecto/serviceinterfaces/IUsuarioService.java
@@ -7,7 +7,9 @@ import java.util.List;
 public interface IUsuarioService {
     public void insert(Usuario usuario);
     public List<Usuario> list();
-    public void delete(int idUsuario);
-    public Usuario listid(int idUsuario);
+    public void delete(Long idUsuario);
+    public Usuario listid(Long idUsuario);
+
+
 
 }
